Remove unnecessary exports in `distributed.communication` and move `wait` & `barrier` (#48396)
* refactor: move wait

* refactor: move barrier

* fix: fix incorrect import
